requirements.txt
PyYAML
python-dotenv
selenium